refactor: pipeline
