+++ httprunner/report.py
@@ -6,14 +6,13 @@ import platform
 import time
 import unittest
 from base64 import b64encode
-from collections import Iterable, OrderedDict
+from collections import Iterable
 from datetime import datetime
 
 from httprunner import logger
 from httprunner.__about__ import __version__
 from httprunner.compat import basestring, bytes, json, numeric_types
 from jinja2 import Template, escape
-from requests.structures import CaseInsensitiveDict
 
 
 def get_platform():
@@ -25,6 +24,7 @@ def get_platform():
         ),
         "platform": platform.platform()
     }
+
 
 def get_summary(result):
     """ get summary from test result
@@ -57,6 +57,25 @@ def get_summary(result):
         summary["records"] = []
 
     return summary
+
+
+def aggregate_stat(origin_stat, new_stat):
+    """ aggregate new_stat to origin_stat.
+
+    Args:
+        origin_stat (dict): origin stat dict, will be updated with new_stat dict.
+        new_stat (dict): new stat dict.
+
+    """
+    for key in new_stat:
+        if key not in origin_stat:
+            origin_stat[key] = new_stat[key]
+        elif key == "start_at":
+            # start datetime
+            origin_stat[key] = min(origin_stat[key], new_stat[key])
+        else:
+            origin_stat[key] += new_stat[key]
+
 
 def render_html_report(summary, html_report_name=None, html_report_template=None):
     """ render html report with specified report name and template
@@ -112,6 +131,7 @@ def render_html_report(summary, html_report_name=None, html_report_template=None
 
     return report_path
 
+
 def stringify_data(meta_data, request_or_response):
     """
     meta_data = {
@@ -151,6 +171,7 @@ def stringify_data(meta_data, request_or_response):
 
         meta_data[request_or_response][key] = value
 
+
 class HtmlTestResult(unittest.TextTestResult):
     """A html result class that can generate formatted html results.
 
@@ -161,12 +182,16 @@ class HtmlTestResult(unittest.TextTestResult):
         self.records = []
 
     def _record_test(self, test, status, attachment=''):
-        self.records.append({
+        data = {
             'name': test.shortDescription(),
             'status': status,
             'attachment': attachment,
-            "meta_data": test.meta_data
-        })
+            "meta_data": {}
+        }
+        if hasattr(test, "meta_data"):
+            data["meta_data"] = test.meta_data
+
+        self.records.append(data)
 
     def startTestRun(self):
         self.start_at = time.time()
