+++ tests/data/demo_parameters.yml
@@ -8,6 +8,7 @@
     variables:
         - device_sn: ${gen_random_string(15)}
         - os_platform: 'ios'
+        - app_version: 2.8.5
     request:
         base_url: $BASE_URL
         headers:
@@ -18,21 +19,9 @@
 
 - test:
     name: get token with $user_agent and $app_version
-    parameters:
-        - app_version: ${gen_app_version()}
     api: get_token($user_agent, $device_sn, $os_platform, $app_version)
     extract:
         - token: content.token
     validate:
         - "eq": ["status_code", 200]
         - "len_eq": ["content.token", 16]
-
-# - test:
-#     name: create user
-#     parameters:
-#         - user_id: [1001, 1002, 1003]
-#         - username-password: ${P(account.csv)}
-#     api: create_user($user_id, $username, $password, $token)
-#     validate:
-#         - {"check": "status_code", "expect": 201}
-#         - {"check": "content.success", "expect": true}
